Working on the actual dictionary lowering code.
I'm learning now why the old code had that separate `Expression` type
for carrying an Expr + a type - it was a way to make every single
expression typed (I guess without having to duplicate a `type` field to
each variant?). Some of the code (marked with a TODO) depends on being
able to read the type of the expression at the point where a class is
being used. Probably this is necessary for selecting the right
implementation of the class in some cases.

There's also just a lot more code left to port (e.g. all of the
lookupDictionary stuff).

I think also that some unit tests are in order for sanity-checking
that all of this works.